Refactored prompts into the gui package.
